integrates audio functionality into the new answer view, adds javascript into asset pipeline, adds css to answers.scss, adds workers to public/lib
