+++ app/assets/javascripts/timer.js
