+++ app/assets/stylesheets/application.css
@@ -9,9 +9,9 @@
  * compiled file so the styles you add here take precedence over styles defined in any styles
  * defined in the other CSS/SCSS files in this directory. It is generally better to create a new
  * file per style scope.
- *
- *= require_tree .
+ *= require answers
  *= require_self
+ *= require_tree .
  */
 
 * {
