+++ public/lib/libopus.js
