+++ public/lib/recorderWorker.js
