--- sandbox/audio-test.html
@@ -1,91 +1,0 @@
-<!DOCTYPE HTML>
-<html>
-  <title>Test</title>
-  <style>
-    button {
-      width: 30px;
-      height: 30px;
-    }
-    audio {
-      margin-left: 200px;
-    }
-    .record-inactive {
-      background-color: red;
-      border-radius: 50px;
-    }
-    .record-active {
-      background-color: grey;
-      border-radius: 0px;
-    }
-    .record-paused {
-      background-color: blue;
-      border-radius: 0px;
-    }
-    .record:active, .record:focus {
-      outline: 0;
-      border:0;
-    }
-    .play {
-      margin-top: 20px;
-    }
-    .play-inactive {
-      background-color: transparent;
-      width: 0; 
-      height: 0; 
-      border-top: 15px solid transparent;
-      border-bottom: 15px solid transparent;
-      border-left: 30px solid green;
-    }
-    .play-active {
-      width: 30px;
-      height: 30px;
-      background-color: red;
-    }
-    .play-paused {
-      width: 30px;
-      height: 30px;
-      background-color: blue
-    }
-    .playback-text{
-      color: green;
-    }
-    .code-playback{
-      background-color: #E6F5EB;
-    }
-    textarea {
-      display: block;
-    }
-    .code-area {
-      width: 500px;
-      height: 700px;
-    }
-    .code-disabled {
-      background-color: lightgrey;
-    }
-    .hide {
-      display: none;
-    }
-
-  </style>
-  <body>
-
-  <h1 id="timer">&nbsp;</h1>
-  <h2 id="question" class="hide">Describe how GitHub optimizes your workflow</h2>
-    <button id="record-btn" class="record record-inactive"></button>
-    <div id="play-btn" class="play play-inactive hide firstplay"></div>
-    <audio controls id="audio" type="audio/ogg">Audio recording not supported on your browser, switch to Firefox or Chrome</audio>
-
-    <textarea id="code-area" class="code-area code-disabled" disabled>
-      
-    </textarea>
-
-    <script type="text/javascript" src="/lib/modernizr.js"></script>
-    <script type="text/javascript" src="/lib/recorderWorker.js"></script>
-    <script type="text/javascript" src="/lib/recorder.js"></script>
-    <script type="text/javascript" src="/audio-test.js"></script>
-    <script type="text/javascript" src="/form-test.js"></script>
-    <script type="text/javascript" src="/timer-test.js"></script>
-    <script type="text/javascript" src="/master-test.js"></script>
-
-  </body>
-</html>
