--- sandbox/server.rb
@@ -1,7 +1,0 @@
-require 'sinatra'
-
-set :root, File.dirname(__FILE__)
-
-get '/' do
-  File.read('audio-test.html')
-end
